move util tests to their own file
